fix(billing): contractDays is int — cast to double in contract-days getters

+++ app/lib/view/station_cleaning/billing/area_rate_config_screen.dart
@@ -111,7 +111,7 @@ class _AreaRateConfigScreenState extends State<AreaRateConfigScreen> {
 
   double get _annualContractValue => _config?.annualContractValue ?? 0;
 
-  double get _contractDays => (_config?.contractDays ?? 0) > 0 ? _config!.contractDays : 365;
+  double get _contractDays => (_config?.contractDays ?? 0) > 0 ? _config!.contractDays.toDouble() : 365;
 
   double get _dailyContractValue => _annualContractValue / _contractDays;
